Split admin job audit server helpers
